Replace root action
The root action was a cohort-extractor action; now it's an sqlrunner
action. We don't need a study definition, but we do need a query. We've
short-circuited the sqlrunner action to write dummy data; this is for
testing the project pipeline. Indeed, if you run `opensafely run
run_all`, then the sqlrunner action should write *input/rows.csv* to
*output/rows.csv*.