Add check job
Eventually, this workflow will resemble a tech-style workflow, rather
than a research-style workflow, as we will have rewrite the test job (we
can't pass a dummy data file to the SQL Runner action when run by CI but
not when run by job-server). This commit starts the transformation by
renaming the workflow, and by adding the check job.