Rewrite test job
This continues the transformation from a research-style workflow to a
tech-style workflow. We can't pass a dummy data file to the SQL Runner
action when run by CI but not when run by job-server. However, if we
don't pass a dummy data file, then the pipeline will fail when run by
CI. So, we rewrite the test job: rather than running the pipeline, we
run the tests.