Add "distinct_values" namespace
We move the query and wrangle actions to the "distinct_values"
namespace. We use this term loosely, meaning a naming convention that
puts similar things in the same group. Here, we use the
"distinct_values" namespace to put inputs (dummy data, queries, and
scripts), outputs (rows, results), and actions (in *project.yaml*) in
the same group, just as we did for the "date_range" namespace.