Count distinct `<Appointment_ID, Organisation_ID>`
Previously, we counted distinct `<Appointment_ID>`. However, the number
of distinct values didn't equal the number of values. Now, we count
distinct `<Appointment_ID, Organisation_ID>`.

We update the query, being mindful to preserve the data structure.
(Consequently, we need not update the dummy data file.) We also refactor
the wrangle action, to ensure a variable is named appropriately. If we
end up doing this again, we may wish to remove the reference to the
tuple from the notebook.