Reduce the risk of disclosing small values
*Small values* are values less than `multiple`. Previously, these values
were rounded to zero. However, group by queries seldom return zeros, so
a rounded value of zero indicated an unrounded value of between one and
`multiple - 1`.

If small values are rounded to `multiple`, rather than to zero, then a
rounded value of `multiple` indicates a much larger range: for example,
if `multiple = 7`, then the range is from `0` to `10` (rather than from
`1` to `6`). It also preserves useful information: that there was an
unrounded value. This reduces the risk of disclosing small values.