Handle all missing dates within the range
Previously, missing dates within the range were handled in a
data-dependant way. If a column had a date value for a given date, then
the other column would have an NaN value for that date. However, if
neither column had a date value for a given date, then neither column
would have an NaN value for that date: the given date would not be
represented, in other words, and Matplotlib would interpolate between
the dates before and after the given date. Although unlikely, this is
undesirable, so we handle all missing dates within the range.