Replace `wrangle_distinct_values`
The `wrangle_distinct_values` action is for rounding and for deriving a
single value. We have a generic rounding action and the derivation of a
single value hasn't been especially useful: it tells us that two counts
are the same, rather than within seven units of each other, which isn't
important in the context of 1.8 billion records. Consequently, we
replace `wrangle_distinct_values` with `round_distinct_values`.