Count distinct values by Organisation_ID
In effect, this commit reverts 22bd7f2 and 57760e7. However, it updates
the notebook, retaining some information and removing the call to
`pandas.read_csv`. This makes it easier to establish whether rows in the
`Appointment` table represent distinct appointments in some cases but
not others, as we need edit only the query.