Disable nbconvert timeout
The most recent `make_html_reports` job failed because of a
`TimeoutError`. #32 added four cells to *analysis/reports/report.ipynb*,
so we conclude that (at least) one of these is taking longer than the
default 30s to execute. We could guess at a suitable value > 30s, but
it's unlikely that the cell will take prohibitively longer, so instead
we disable the timeout.