Count rows by `Status`
Previous results tell us that there are more unique values of
`Appointment.Status` than are present in `DataDictionary.Code`. Whether
these unique values are significant depends on how many rows in
`Appointment` are associated with them. So, we count the rows and,
because it's hard to relate to a number (`Status`/`Code`), we join to
`DataDictionary` to access `DataDictionary.Description`.